Rename graphql/helpers -> graphql/utils for consistency
