+++ src/mirage/resolvers/object.ts
@@ -1,6 +1,6 @@
 import { Resolver, ResolverInfo, ResolverParent } from '../../types';
 import { MirageGraphQLMapper } from '../mapper';
-import { relayPaginateNodes } from '../../relay/helpers';
+import { relayPaginateNodes } from '../../relay/utils';
 import { mirageCursorForNode, ObjectResolverMatch, AutoResolverErrorMeta } from './helpers';
 import { extractDependencies } from '../../resolver/extract-dependencies';
 import { AutoResolverError } from './auto-resolver-error';
