+++ test/integration/test-helpers/mirage-static-resolvers/Query/all-persons-paginated.ts
@@ -1,9 +1,10 @@
 import { extractDependencies } from '../../../../../src/resolver/extract-dependencies';
-import { relayPaginateNodes } from '../../../../../src/relay/helpers';
+import { relayPaginateNodes } from '../../../../../src/relay/utils';
 import { Server } from 'miragejs';
 
 export default function (_parent: unknown, args: Record<string, unknown>, context: Record<string, unknown>): unknown {
   const { mirageServer } = extractDependencies<{ mirageServer: Server }>(['mirageServer'], context);
   const nodes = mirageServer.schema.all('person').models;
-  return relayPaginateNodes<Record<string, unknown>>(nodes, args, (node) => node.toString());
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  return relayPaginateNodes<Record<string, unknown>>(nodes, args, (node: any) => node.toString());
 }
